refator: move to class based structure
