+++ pdda/__init__.py
@@ -1,17 +1,11 @@
-from pdda.augment.augmenter import SingleSigAugmenter, DualSignalAugmenter
-from pdda.augment.st_aug import emd_augmentation, mix_augmentation
-from pdda.augment.fr_aug import frequency_mixing, frequency_masking
-from pdda.augment.wave_aug import wave_mix, wave_mask
-from pdda.augment.d_shuffle import dominant_shuffle
+from pdda.augments.wave_aug import WaveAug
+from pdda.augments.fr_aug import FRAug
+from pdda.augments.st_aug import STAug
+from pdda.augments.d_shuffle import DShuffle
 
 __all__ = [
-    "SingleSigAugmenter",
-    "DualSignalAugmenter",
-    "emd_augmentation",
-    "mix_augmentation",
-    "frequency_mixing",
-    "frequency_masking",
-    "wave_mix",
-    "wave_mask",
-    "dominant_shuffle",
+    "WaveAug",
+    "FRAug",
+    "STAug",
+    "DShuffle",
 ]
